+++ lib/stores/consumerProgressStore/MySql/MySqlConsumerProgressStore.ts
@@ -1,6 +1,6 @@
 import { AggregateIdentifier } from '../../../common/elements/AggregateIdentifier';
 import { ConsumerProgressStore } from '../ConsumerProgressStore';
-import { createPoolWithDefaults } from '../../utils/mySql/createPool';
+import { createPoolWithDefaults } from '../../utils/mySql/createPoolWithDefaults';
 import { errors } from '../../../common/errors';
 import { getHash } from '../../../common/utils/crypto/getHash';
 import { IsReplaying } from '../IsReplaying';
